# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  min_bid?: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AuthInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  userId: number;
  username: string;
  coins: number;
}

export interface PlayerCard {
  id: number;
  inventory_id: number;
  name: string;
  initials: string;
  position: string;
  country: string;
  club: string;
  is_gk: boolean;
  /** @nullable */
  fk?: number | null;
  /** @nullable */
  pk?: number | null;
  /** @nullable */
  fks?: number | null;
  /** @nullable */
  pks?: number | null;
  overall: number;
  tactical_position: string;
  is_bench_pool: boolean;
  is_listed: boolean;
  acquired_at: string;
}

export interface RegistrationResult {
  token: string;
  userId: number;
  username: string;
  coins: number;
  message: string;
  starter_pack: PlayerCard[];
}

export interface DailyReward {
  awarded: number;
  coins: number;
}

export interface BalanceResponse {
  coins: number;
}

export interface InventorySummary {
  coins: number;
  total_cards: number;
  goalkeepers: number;
  shooters: number;
}

export interface ExchangeResult {
  success: boolean;
  coins_earned: number;
  player_name: string;
  new_balance: number;
}

export interface Pack {
  id: string;
  name: string;
  description: string;
  cost: number;
  card_count: number;
}

export interface BuyPackInput {
  packType: string;
}

export interface PackOpenResult {
  players: PlayerCard[];
  remaining_coins: number;
  duplicate_inventory_ids: number[];
}

export interface NerfPreview {
  player_id: number;
  name: string;
  original_role: string;
  used_as_role: string;
  /** @nullable */
  effective_fk: number | null;
  /** @nullable */
  effective_pk: number | null;
}

export interface SquadSlotPlayer {
  inventory_id: number;
  player_id: number;
  name: string;
  initials: string;
  overall: number;
  tactical_position: string;
  is_gk: boolean;
  /** @nullable */
  fk?: number | null;
  /** @nullable */
  pk?: number | null;
  /** @nullable */
  fks?: number | null;
  /** @nullable */
  pks?: number | null;
  club: string;
  country: string;
  is_bench_pool?: boolean;
}

export type SquadSlots = {[key: string]: SquadSlotPlayer | null};

export interface Squad {
  squad_type: string;
  squad_id: number;
  /** @nullable */
  team_ovr?: number | null;
  slots: SquadSlots;
}

export interface SetSlotInput {
  slot_key: string;
  inventory_id: number;
}

export interface MarketListing {
  id: number;
  seller_id: number;
  seller_username: string;
  inventory_id: number;
  player_id: number;
  player_name: string;
  player_initials: string;
  player_overall: number;
  player_position: string;
  player_club: string;
  player_country: string;
  player_is_gk: number;
  starting_bid: number;
  current_bid: number;
  /** @nullable */
  current_bidder_id?: number | null;
  /** @nullable */
  current_bidder_username?: string | null;
  expires_at: number;
  created_at: string;
  quick_sell_value: number;
}

export interface MarketPage {
  listings: MarketListing[];
  page: number;
  total: number;
  pages: number;
}

export type CreateListingInputDuration = typeof CreateListingInputDuration[keyof typeof CreateListingInputDuration];


export const CreateListingInputDuration = {
  '1h': '1h',
  '2h': '2h',
  '6h': '6h',
  '12h': '12h',
  '24h': '24h',
} as const;

export interface CreateListingInput {
  inventory_id: number;
  starting_bid: number;
  duration: CreateListingInputDuration;
}

export interface PlaceBidInput {
  amount: number;
}

export type DailyObjectiveSquadSlotsItem = { [key: string]: unknown };

export type DailyObjectiveSquad = {
  slots?: DailyObjectiveSquadSlotsItem[];
  /** @nullable */
  team_ovr?: number | null;
};

export interface DailyObjective {
  date: string;
  formation: string;
  next_reset_at: string;
  formations_pool: string[];
  squad: DailyObjectiveSquad;
}

export type GetNerfPreviewParams = {
playerId: number;
};

export type GetMarketParams = {
page?: number;
position?: string;
};