# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: FreeKick Gacha Card Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account (auto-grants Free 11-Player Starter Pack)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/daily-claim:
    post:
      operationId: dailyClaim
      tags: [auth]
      summary: Claim daily coin reward (200 coins, once per day)
      responses:
        "200":
          description: Coins awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReward"
        "400":
          description: Already claimed today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        min_bid:
          type: integer
...
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, userId, username, coins]
      properties:
        token:
          type: string
        userId:
          type: integer
        username:
          type: string
        coins:
          type: integer
    RegistrationResult:
      type: object
      required: [token, userId, username, coins, message, starter_pack]
      properties:
        token:
          type: string
        userId:
          type: integer
        username:
          type: string
        coins:
          type: integer
        message:
          type: string
        starter_pack:
          type: array
          items:
            $ref: "#/components/schemas/PlayerCard"
    DailyReward:
      type: object
      required: [awarded, coins]
      properties:
        awarded:
          type: integer
        coins:
          type: integer
...
    PlayerCard:
      type: object
      required: [id, inventory_id, name, initials, position, country, club, is_gk, overall, tactical_position, is_bench_pool, is_listed, acquired_at]
      properties:
        id:
          type: integer
        inventory_id:
          type: integer
        name:
          type: string
        initials:
          type: string
        position:
          type: string
        country:
          type: string
        club:
          type: string
        is_gk:
          type: boolean
        fk:
          type: ["integer", "null"]
        pk:
          type: ["integer", "null"]
        fks:
          type: ["integer", "null"]
        pks:
          type: ["integer", "null"]
        overall:
          type: integer
        tactical_position:
          type: string
        is_bench_pool:
          type: boolean
        is_listed:
          type: boolean
        acquired_at:
          type: string
...