# lib/api-zod/src/generated/types/playerCard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */

export interface PlayerCard {
  id: number;
  inventory_id: number;
  name: string;
  initials: string;
  position: string;
  country: string;
  club: string;
  is_gk: boolean;
  /** @nullable */
  fk?: number | null;
  /** @nullable */
  pk?: number | null;
  /** @nullable */
  fks?: number | null;
  /** @nullable */
  pks?: number | null;
  overall: number;
  tactical_position: string;
  is_bench_pool: boolean;
  is_listed: boolean;
  acquired_at: string;
}