# lib/api-zod/src/generated/types/registrationResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */
import type { PlayerCard } from './playerCard';

export interface RegistrationResult {
  token: string;
  userId: number;
  username: string;
  coins: number;
  message: string;
  starter_pack: PlayerCard[];
}